Amazon process updates
* Archive old donateAmazon function
* Refactor funciton to hide form on returning from Amazon
