+++ donatenow/js/luminateExtend-custom.js
@@ -309,13 +309,16 @@ function submitToVenmo() {
 	braintree_aha.submitVenmoDonation();
 }
 
-function donateAmazon(amazonCheckoutSessionId) {
+if (location.href.indexOf("amazonCheckoutSessionId") > 0) {
+	// hide form - show loading
 	window.scrollTo(0, 0);
-	$('.donation-form').hide();
-	$('.donation-form').before('<div class="well donation-loading">' +
+	jQuery('.donation-form').hide();
+	jQuery('.donation-form').before('<div class="well donation-loading">' +
 			'Thank You!  We are now processing your donation from Amazon ...' +
 			'</div>');
+}
 
+function donateAmazon(amazonCheckoutSessionId) {
 	let lsForm = localStorage.getItem('ahaDonate');
 	if (lsForm != null) {
 		// verify checkout
@@ -330,122 +333,6 @@ function donateAmazon(amazonCheckoutSessionId) {
 		$('.donation-form').show();
 	}
 }
-
-function donateAmazonOld() {
-	window.scrollTo(0, 0);
-	$('.donation-form').hide();
-	$('.donation-form').before('<div class="well donation-loading">' + 
-					 'Thank You!  We are now processing your donation from Amazon ...' +
-				   '</div>');
-	// read saved form data
-	// add 
-	var params = $('.donation-form').serialize();
-	var amazonErr = false;
-	var status = "";
-	var amt = 0;
-	var ref = 0;
-
-	$.ajax({
-		method: "POST",
-		async: false,
-		cache:false,
-		dataType: "json",
-		url:"https://tools.heart.org/donate/amazon/payWithAmazon.php?"+params+"&callback=?",
-		success: function(data){
-			if ($('label[for="type-monthly"] .active').length > 0) {
-				status = data.data.AuthorizeOnBillingAgreementResult.AuthorizationDetails.AuthorizationStatus.State;
-				amt = data.data.AuthorizeOnBillingAgreementResult.AuthorizationDetails.CapturedAmount.Amount;
-				ref = data.data.AuthorizeOnBillingAgreementResult.AuthorizationDetails.AmazonAuthorizationId;
-
-				if (status != "Closed") {
-					amazonErr = true;
-				}
-			} else {
-				status = data.data.AuthorizeResult.AuthorizationDetails.AuthorizationStatus.State;
-				amt = data.data.AuthorizeResult.AuthorizationDetails.CapturedAmount.Amount;
-				ref = data.data.AuthorizeResult.AuthorizationDetails.AmazonAuthorizationId;
-
-				if (status != "Closed") {
-					amazonErr = true;
-				}
-			}
-
-			if (amazonErr) {
-				$('#donation-errors').append('<div class="alert alert-danger">' + data.data.toString() + '</div>');	
-
-				$('.donation-loading').remove();
-				$('.donation-form').show();				
-			} else {
-				//save off amazon id into custom field
-				$('input[name=check_number]').val(ref);
-				$('input[name=payment_confirmation_id]').val('AMAZON:'+ref);
-
-				//logout of amazon
-				amazon.Login.logout();
-
-				//make offline donation in luminate to record transaction
-				// if ($('input[name="df_preview"]').val() != "true") donateOffline();
-				donateOffline(donateOfflineCallback);
-
-				var email = $('input[name="donor.email"]').val();
-				var first = $('input[name="donor.name.first"]').val();
-				var last = $('input[name="donor.name.last"]').val();
-				var street1 = $('input[name="donor.address.street1"]').val();
-				var street2 = $('input[name="donor.address.street2"]').val();
-				var city = $('input[name="donor.address.city"]').val();
-				var state = $('[name="donor.address.state"]').val();
-				var zip = $('input[name="donor.address.zip"]').val();
-				var country = $('select[name="donor.address.country"]').val();
-				var form=$('input[name=form_id]').val();
-
-			  $('.donation-loading').remove();
-			  $('.donate-now, .header-donate').hide();
-			  $('.thank-you').show();
-			  $.get(donation_thank_you_page,function(datat){ 
-				  $('.thank-you').html($(datat).find('.thank-you').html());
-				  $('p.first').html(first);
-				  $('p.last').html(last);
-				  $('p.street1').html(street1);
-				  $('p.street2').html(street2);
-				  $('p.city').html(city);
-				  $('p.state').html(state);
-				  $('p.zip').html(zip);
-				  $('p.country').html(country);
-				  $('p.email').html(email);
-				  $('tr.cardGroup').hide();
-				  $('tr.amazon').show();
-				  $('p.amount').html("$"+amt);
-				  $('p.confcode').html(ref);
-				});
-
-				$('.thank-you').append('<img src="//offeredby.net/silver/track/rvm.cfm?cid=28556&oid='+ref+'&amount='+amt+'&quantity=1" height="1" width="1">');
-				$.getScript("//action.dstillery.com/orbserv/nsjs?adv=cl1014039&ns=1985&nc=HBP-Donate-Now-Landing-Page&ncv=52&dstOrderId="+ref+"&dstOrderAmount="+amt);
-
-				// Custom Tracking code
-				includeCustomFBPixel(amt);
-				
-				/* ECOMMERCE TRACKING CODE */ 
-				ga('require', 'ecommerce');
-
-				ga('ecommerce:addTransaction', {
-				  'id': ref,
-				  'affiliation': 'AHA Amazon Donation',
-				  'revenue': amt,
-				  'city': $('input[name="donor.address.city"]').val(),
-				  'state': $('select[name="donor.address.state"]').val()  // local currency code.
-				});
-
-				ga('ecommerce:send');
-
-				ga('send', 'pageview', '/donateok.asp');
-				
-				pushDonationSuccessToDataLayer(form, ref, amt);
-			}
-		}
-	});
-
-}
-
 
 function donateApplePay() {
 	window.scrollTo(0, 0);
@@ -1032,9 +919,11 @@ $('[name^=donor\\.]').each(function(){
 
 	// Get Amazon confirmation id
 	const amzConfirmationId = $.getQuerystring('amazonCheckoutSessionId');
-	if (amzConfirmationId) {
-		donateAmazon(amzConfirmationId);
-	}
+	jQuery(document).ready(function(){
+		if (amzConfirmationId) {
+			donateAmazon(amzConfirmationId);
+		}
+	});
 
 // END QUERY STRING CODE
